edit the A_input(), A_output(), A_timerinterrupt(), A_init()
I had edit A_input so that it handles Acks, slide window. Edit A_output to send packet if window is not full and stores them in the buffer. I aslow edit the A_timerinterrupt() so that it resends all unacked packed. I also replace most of the A_nexteqnum with just nextqnum since was used for gbn.